Stop using the workaround for bug #17329
- Update the spec in extensions/:
There was a bug in dbus-glib that prevent to use the right type:
Instead of a{ua(a{sv}as)}, we used a(ua{sv}as) as a workaround.
See http://bugs.freedesktop.org/show_bug.cgi?id=17329
Now there is a fix, so we don't use the workaround anymore.

- Update the implementation for the new spec

- Update the unit test

This commit is similar to the following commit in telepathy-gabble:
|commit 6971f458ad0816e40807549202d8a4ffe58b72aa
|Author: Alban Crequy <alban.crequy@collabora.co.uk>
|Date:   Wed Oct 8 11:30:59 2008 +0100