Don't announce Tubes channel twice when requesting a StreamTube
Previously, when a Channel.Type.Tubes was created as a side-effect of
requesting a Channel.Type.StreamTube, the implementation attempted to
emit a single NewChannels signal for both channels. But it was
incorrect: it emitted new-channels([Tubes], requests_satisfied=[])
followed by new-channels([Tubes, StreamTube], requests_satisfied=[the
request you first thought of]) because passing request_token=NULL to
new_tubes_channel() does not stop it emitting NewChannels.

On the assumption that no-one actually cares that the Tubes channel is a
side-effect of the StreamTube channel, let's just announce them
separately and be done with it.