Unconfuse TpTubeChannelState and TpTubeState a bit
My version of GCC didn't like building a bunch of the tube code because
it confuses TpTubeChannelState and TpTubeState. In practice this code
works because the former is a superset of the latter and none of the
confused code paths can encounter the extra value. This patch tidies up
just enough for the compiler not to whine.