plugin-loader: don't critical if a plugin lazily implements no sidecars
An empty strv is really { NULL }, but it's not so unreasonable to
use just NULL instead, so we shouldn't break if we get that.

Creating a sidecar is safe as it checks whether the strv contains the
requested interface using tp_strv_contains() which treats NULL as an
empty strv.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>