check-gibber-listener: fix test by not depending on what test was run before
The main loop (and therefore the listener) was only started if the
signalled gboolean was FALSE. If this was the first test run then this
would be the case, but if the tcp-listen test was run beforehand then
it would be TRUE and so the main loop wouldn't run and the test would
fail.

Seeing that the unix-listen test doesn't actually ever test the
signalled gboolean, let's just stop depending on it.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>