util: don't try and get caps out of something we can't read
Most link-local contacts in the contact factory will be created by the
avahi contact manager in Salut, and so they will ultimately be
SalutAvahiContact objects, but not all of them.

One example of one which isn't is the LL contact which wocky ensures
when creating the loopback porter in WockyMetaPorter, so let's not
cast that to a SalutContact blindly.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>