file_transfer_helper: use a predicate to find the FT channel
Picking up the first channel that pops up from NewChannels can be
wrong. There's something up with the tests on the buildbot but not
sure what -- this might fix something there though. The worst this can
do is change the test failure from an AssertionError to a
TimeoutError which is fine.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>