tests: look for 'contact-ids' in MembersChangedDetailed
For at least a year, telepathy-glib used the wrong name for this field
in the details of a group changed: member-ids, rather than contact-ids
as defined in the spec and used by TelepathyQt4.

Since Salut doesn't depend on tp-glib >= 0.14.6, let's look for both in
the test suite.

Looked-over-my-shoulder-while-I-fixed-this: Jonny Lamb <jonny.lamb@collabora.co.uk>