Update libglibcodegen from telepathy-glib master, and add libtpcodegen
This requires the imports in glib-ginterface-gen to be updated too.
Sync the entire file. It no longer needs signals-marshal.h to be
generated, so drop that.

This also introduces more strict checks for extension interfaces,
so sync the OLPC interfaces from Gabble master, and add Room_Handle to
the list of external types in all.xml.