Add extensions to symbol-hacks
Because the test interface is the last thing in the extensions static
library, we have to add it to symbol-hacks.c so that it's visible to
the plugin. This wouldn't work on platforms where plugins can't have
undefined symbols, like Windows - we'd need something more like
Mission Control's mission-control-plugins or Gabble's libgabble-plugins
there - but it'll do for now.